% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{Aijfcn}
\alias{Aijfcn}
\title{A function to calculate the nitrogen surplus or deficit each species gets from consuming another species}
\usage{
Aijfcn(usin)
}
\arguments{
\item{usin}{The community on which nitrogen calculations are made.}
}
\value{
A matrix of nitrogen surpluses or deficits.
}
\description{
A function to calculate the nitrogen surplus or deficit each species gets from consuming another species
}
\examples{
Aijfcn(intro_comm)
}
