% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{Andres2016}
\alias{Andres2016}
\title{The soil food webs published for grazed and ungrazed plots in the Shortgrass Steppe long-term research station.}
\format{
A list of six communities each with a feeding matrix and properties dataframe. NOTE: You must select one of the communities from the list to use the package functions. For example: comana(Andres2016$GA) carries out calculations for the GA plot.
\describe{
  \item{GA}{The grazed plot A.}
  \item{UGA}{The ungrazed plot A.}
  \item{GB}{The grazed plot B.}
  \item{UGB}{The ungrazed plot B.}
  \item{GC}{The grazed plot C.}
  \item{UGC}{The ungrazed plot C.}
  \item{imat}{Within the community: The feeding matrix. Rows eat columns.}
  \item{prop}{Within the community: The properties data frame containing node names (ID), assimilation efficiency (a), production efficiency(p), C:N ratio (CN), biomass (B), death rate (d), proportion of death cycled back to a detrital pool (DetritusRecycling), Booleans stating whether the node is detritus, plant, and can immobilize nitrogen, and a list of mutual predators. Biomass is in kilograms of carbon per hectare and turnover/death rate is in years.}
}
}
\source{
\doi{10.1016/j.soilbio.2016.02.014}
}
\usage{
Andres2016
}
\description{
The community contains 21 nodes in bacterial, fungal, and root energy channels.
}
\keyword{datasets}
