% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_foodweb.R
\name{build_foodweb}
\alias{build_foodweb}
\title{A function to compile the food web from simple data inputs}
\usage{
build_foodweb(feeding, properties)
}
\arguments{
\item{feeding}{A data frame listing the feeding relationships in the food web. Only needs to contain the nodes that have prey items along with their names. Three columns in this data frame must be: Predator, Prey, and Preference. The Predator column is the name of the predator, the Prey column is the name of the prey, and the Preference is the preference that the predator has for that prey item after correcting for abundance. Preference values are relative, so for each predator the preference value can be any number and all that matters is its proportion of the total of all preference values given to that predator. Preference is meaningless if a predator only has one prey item. A good default value is 1 for everything if you don't want to set preferences beyond prey abundance.}

\item{properties}{A data frame listing the properties or parameters in the food web. Must contain the names of all of the nodes in the food web. This data frame must contain: ID, d,a,p,B, CN, Detritusrecycling, isDetritus, isPlant, and canIMM.}
}
\value{
A community that is compatible with the functions of soilfoodwebs. This is a list with the feeding matrix imat first and the properties prop second.
}
\description{
A function to compile the food web from simple data inputs
}
