% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_smin.R
\name{calc_smin}
\alias{calc_smin}
\title{Calculate the strength of stability as smin.}
\usage{
calc_smin(usin)
}
\arguments{
\item{usin}{The community for which to calculate smin}
}
\value{
The values of smin.
}
\description{
Calculate the strength of stability as smin.
}
\details{
Calculates the value of smin in each web using the metric used by Moore and de Ruiter.
}
\examples{
# Calculate the minimum values of s for the introduction community.
calc_smin(intro_comm)
}
