% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_smin.R
\name{calculate_smin}
\alias{calculate_smin}
\title{Function used inside the calc_smin function}
\usage{
calculate_smin(SMIN = 1, usin, isSQR = TRUE)
}
\arguments{
\item{SMIN}{The value of smin.}

\item{usin}{The community used in the calculation}

\item{isSQR}{Boolean: Should the rmax value be squared?}
}
\value{
The value of rmax for the community and given SMIN
}
\description{
Function used inside the calc_smin function
}
