% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{can_mutfeed}
\alias{can_mutfeed}
\title{A function that identifies cannibalism and mutual feeding}
\usage{
can_mutfeed(usin)
}
\arguments{
\item{usin}{The community in which to identify cannibalism and mutual feeding}
}
\value{
The community with mutual feeding added to the properties database.
}
\description{
A function that identifies cannibalism and mutual feeding
}
