% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{checkeqm}
\alias{checkeqm}
\title{Check the carbon flux equilibrium output by comana.}
\usage{
checkeqm(cares)
}
\arguments{
\item{cares}{The output from comana.}
}
\value{
Boolean: Is the community at equilibrium with inputs to the first trophic level?
}
\description{
Check the carbon flux equilibrium output by comana.
}
\examples{
checkeqm(comana(intro_comm))
}
