% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{Cijfcn}
\alias{Cijfcn}
\title{A utility function to calculate the consumption rate of each species on all prey assuming a type I functional response.}
\usage{
Cijfcn(usin, shuffleTL = FALSE, rmzeros = TRUE)
}
\arguments{
\item{usin}{The community on which feeding rate calculations are made.}

\item{shuffleTL}{A Boolean stating whether the community should be sorted before consumption rates are calculated.}

\item{rmzeros}{A Boolean determining whether trophic species with zero biomass should be removed from the community.}
}
\value{
A matrix of consumption rates with units set by the the biomass input units in biomass and time.
}
\description{
A utility function to calculate the consumption rate of each species on all prey assuming a type I functional response.
}
\examples{
Cijfcn(intro_comm)
}
