% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtrosp.R
\name{comtrosp_backend}
\alias{comtrosp_backend}
\title{A function to combine trophic species with only two seltf options available}
\usage{
comtrosp_backend(
  usin,
  seltf = F,
  selected = NA,
  deleteCOMBOcannibal = F,
  allFEEDING1 = F
)
}
\arguments{
\item{usin, }{# The community were you are combining trophic species}

\item{seltf}{Boolean: Do you want to select the trophic species to combine or do you want to combine two based on their similarity?}

\item{selected}{A vector of two trophic species to combine. The default is NA, which is not appropriate if seltf is T.}

\item{deleteCOMBOcannibal}{Boolean: Do you want to delete the cannibalism that may have been created by combining two trophic species (T) or leave it in the model (F)?}

\item{allFEEDING1}{Boolean: Do you want to return all feeding preferences to 1 (T), or would you like to set the feeding preferences of the newly combined trophic species as the biomass-weighted average of the old ones (F)?}
}
\value{
The new community with the seltf or most similar trophic species combined.
}
\description{
A function to combine trophic species with only two seltf options available
}
