% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrstoich.R
\name{correct_diet}
\alias{correct_diet}
\title{A function to correct the diet of trophic species.}
\usage{
correct_diet(usin, dietlimits = c(NA))
}
\arguments{
\item{usin}{The input community in which to fix the diets.}

\item{dietlimits}{# A matrix the same size as imat that gives the diet limits as a proportion of the total diet. All values must be between 0 and 1. Leaving it as NA sets the limits of all diet items to 1.}
}
\value{
The modified community with new diet preferences.
}
\description{
A function to correct the diet of trophic species.
}
