% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrstoich.R
\name{corrstoich}
\alias{corrstoich}
\title{A function that corrects trophic species production efficiency and/or diet composition to balance carbon and nitrogen demands.}
\usage{
corrstoich(usin, forceProd = F, dietlimits = c(NA), Immobilizationlimit = Inf)
}
\arguments{
\item{usin}{# Community in which to correct stoichiometry.}

\item{forceProd}{Boolean: Should we force organisms to only change their production efficiency and not their diet?}

\item{dietlimits}{A matrix the same size as imat that gives the diet limits as a proportion of the total diet. All values must be between 0 and 1. Leaving it as NA sets the limits of all diet items to 1.}

\item{Immobilizationlimit}{Set a limit for the maximum rate of immobilization of inorganic nitrogen for the food web per time step.}
}
\value{
The modified community with new production efficiencies and diets.
}
\description{
A function that corrects trophic species production efficiency and/or diet composition to balance carbon and nitrogen demands.
}
