% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompexpt.R
\name{decompexpt}
\alias{decompexpt}
\title{Decomposition rates and effect of individual organisms}
\usage{
decompexpt(usin)
}
\arguments{
\item{usin}{The community in which we want to calculate decomposition rates.}
}
\value{
A table of decomposition constants for each detritus pool.
}
\description{
Decomposition rates and effect of individual organisms
}
\details{
The output list indicates the baseline decomposition rate (k; basedecomp) and then a table showing the effect on decomposition rates when the direct and indirect effects of individual trophic species are removed (decompeffects). Removing direct effects means setting consumption of that species to zero, while indirect effects occur when the community is recalculated without that species before calculating the decompostion rate (k). A positive number means the species increases decomposition. The column ID contains the node having the ffect, while DetritusID indicates the detritus pool whose rate is being reported.
Direct effects are calculated as the decomposition rate in the full community minus decomposition rate without the species consumption of detritus, divided by the decomposition rate in the full community.
Indirect effects are calculated as the difference in decomposition rate of the full community minus decomposition rate without the species and minus, divided by the decomposition rate with the species and minus the direct effect. A negative value here means that removing the species reduces decomposition more than any direct consumption effects it has on detritus.
}
\examples{
# Basic example for the introductory community:
decompexpt(intro_comm)
}
