% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{deRuiter1994}
\alias{deRuiter1994}
\title{The soil food webs published for conventional (CON) and integrated (INT) management at Lovinkhoeve experimental farm.}
\format{
A list of four communities each with a feeding matrix and properties dataframe. NOTE: You must select one of the communities from the list to use the package functions. For example: comana(deRuiter1994$INT) carries out calculations for the integrated field.
\describe{
  \item{INT}{The field with integrated management, 0 to 10-cm.}
  \item{INT10}{The field with integrated management, 10 to 25-cm.}
  \item{CON}{The field with conventional management, 0 to 10-cm.}
  \item{CON10}{The field with conventional management, 10 to 25-cm.}
  \item{imat}{Within the community: The feeding matrix. Rows eat columns.}
  \item{prop}{Within the community: The properties data frame containing node names (ID), assimilation efficiency (a), production efficiency(p), C:N ratio (CN), biomass (B), death rate (d), proportion of death cycled back to a detrital pool (DetritusRecycling), Booleans stating whether the node is detritus, plant, and can immobilize nitrogen, and a list of mutual predators. Biomass is in kilograms of carbon per hectare in the depth range noted above and turnover/death rate is in years.}
}
}
\source{
\doi{10.1016/0167-8809(94)90044-2}
}
\usage{
deRuiter1994
}
\description{
The community contains 18 nodes in bacterial and fungal energy channels.
}
\keyword{datasets}
