% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodwebode.R
\name{foodwebode}
\alias{foodwebode}
\title{A function to simulation the food webs away from equilibrium.}
\usage{
foodwebode(t, y, pars)
}
\arguments{
\item{t}{The ODE time.}

\item{y}{The ODE simulation start.}

\item{pars}{The ODE parameters.}
}
\value{
The changes in each node biomass along with parameters and mineralization rates.
}
\description{
A function to simulation the food webs away from equilibrium.
}
\details{
The food web model ode for simulating the model over time: requires y inputs and parameters from getPARAM function: it adds a detritus decomposition experiment. The only pools with nitrogen values are the detritus, because all others have fixed C:N ratios.
}
