% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_weights.R
\name{geoMEAN}
\alias{geoMEAN}
\title{A function to calculate the geometric mean}
\usage{
geoMEAN(VEC)
}
\arguments{
\item{VEC}{The vector whose mean you want to calculate.}
}
\value{
The geometric mean.
}
\description{
A function to calculate the geometric mean
}
