% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{Hunt1987}
\alias{Hunt1987}
\title{The soil food web published for CPER}
\format{
A community with a feeding matrix and properties dataframe:
\describe{
  \item{imat}{The feeding matrix. Rows eat columns.}
  \item{prop}{The properties data frame containing node names (ID), assimilation efficiency (a), production efficiency(p), C:N ratio (CN), biomass (B), death rate (d), proportion of death cycled back to a detrital pool (DetritusRecycling), Booleans stating whether the node is detritus, plant, and can immobilize nitrogen, and a list of mutual predators. Biomass is in kilograms of carbon per hectare and turnover/death rate is in years.}
}
}
\source{
\doi{10.1007/BF00260580}
}
\usage{
Hunt1987
}
\description{
The community contains 17 nodes in bacterial, fungal, and root energy channels.
}
\keyword{datasets}
