% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{intro_comm}
\alias{intro_comm}
\title{A baseline community for examples}
\format{
A community with a feeding matrix and properties dataframe:
\describe{
  \item{imat}{The feeding matrix. Rows eat columns.}
  \item{prop}{The properties data frame containing node names (ID), assimilation efficiency (a), production efficiency(p), C:N ratio (CN), biomass (B), death rate (d), proportion of death cycled back to a detrital pool (DetritusRecycling), Booleans stating whether the node is detritus, plant, and can immobilize nitrogen, and a list of mutual predators.}
}
}
\source{
Example not based on real empirical data.
}
\usage{
intro_comm
}
\description{
The community contains seven nodes generally meant to represent oribatid mites feeding on microorganisms and detritus.
}
\keyword{datasets}
