% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_weights.R
\name{is_loop}
\alias{is_loop}
\title{A function to identify a loop in a food web vector}
\usage{
is_loop(VEC)
}
\arguments{
\item{VEC}{The vector to check.}
}
\value{
Boolean
}
\description{
A function to identify a loop in a food web vector
}
