% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jacobsindex.R
\name{Jacobsindex}
\alias{Jacobsindex}
\title{Calculate Jacob's index.}
\usage{
Jacobsindex(usin)
}
\arguments{
\item{usin}{The community for which to calculate Jacob's index}
}
\value{
A data frame with columns for predator, prey, and the value of Jacob's index.
}
\description{
Calculate Jacob's index.
}
\details{
Jacob's index is calculated for all predators with more than one prey item. A value of -1 indicates maximum avoidance, 0 indicates no preference, and 1 indicates maximum preference. This index is a more user friendly output than the values in imat, which can be difficult to interpret.
}
\examples{
# Calculate the minimum values of s for the introduction community.
Jacobsindex(intro_comm)
}
