% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_weights.R
\name{loop_weights}
\alias{loop_weights}
\title{A function to calculate loop weights for a matrix}
\usage{
loop_weights(mat, max_length = 100, max_return = 10)
}
\arguments{
\item{mat}{The matrix for which to calculate loop weights.}

\item{max_length}{The maximum length of paths to check.}

\item{max_return}{The number of loop weights per node you want to return.}
}
\value{
The geometric mean.
}
\description{
A function to calculate loop weights for a matrix
}
\details{
This function calculates the loop weights for each of possible loop up to the length max_length. This calculation can take a long time for large matricies, so a counter is provided. The function returns on the top 10 loop weights.
}
