% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{newnode}
\alias{newnode}
\title{Add node to the community}
\usage{
newnode(COMM, newname, prey = NA, predator = NA, newprops)
}
\arguments{
\item{COMM}{The community to which to add nodes.}

\item{newname}{The new node ID.}

\item{prey}{A vector of prey preferences with names.}

\item{predator}{A vector of predators and their preferences with name.}

\item{newprops}{A vector of the new properties with the appropriate names.}
}
\value{
The community with the new node.
}
\description{
Add node to the community
}
\seealso{
\code{\link{removenodes}}
}
