% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_weights.R
\name{path_n_weights}
\alias{path_n_weights}
\title{A function to calculate the loop weights for one node.}
\usage{
path_n_weights(mat, node = 1, max_length = 100)
}
\arguments{
\item{mat}{The matrix for which to calculate loop weights.}

\item{node}{The node for which to calculate loop weights.}

\item{max_length}{The maximum length of paths to check.}
}
\value{
The loop weights for the chosen node.
}
\description{
A function to calculate the loop weights for one node.
}
