% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrstoich.R
\name{productionadj}
\alias{productionadj}
\title{A function to fix production efficiency.}
\usage{
productionadj(usin, immobilizationlimit = Inf)
}
\arguments{
\item{usin}{The input community to fix the production efficiency.}

\item{immobilizationlimit}{Set a limit for the maximum rate of immobilization of inorganic nitrogen for the food web per time step.}
}
\value{
The modified community with new production efficiency.
}
\description{
A function to fix production efficiency.
}
