% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{renamenode}
\alias{renamenode}
\title{Rename a node in a  community.}
\usage{
renamenode(COMM, oldname, newname)
}
\arguments{
\item{COMM}{The community from which to remove nodes.}

\item{oldname}{The node's old name}

\item{newname}{The node's new name}
}
\value{
The community with the new name.
}
\description{
Rename a node in a  community.
}
\examples{
renamenode(intro_comm, oldname = "Predator", newname = "NewPredator")
}
