% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{RESCALE}
\alias{RESCALE}
\title{A function to rescale a vector.}
\usage{
RESCALE(invec, a = 0, b = 1)
}
\arguments{
\item{invec}{The vector to re-scale.}

\item{a}{The lower limit of the new scale.}

\item{b}{The upper limit of the new scale.}
}
\value{
The scaled vector.
}
\description{
A function to rescale a vector.
}
