% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{TLsort}
\alias{TLsort}
\title{Sorts the trophic levels from lowest to highest}
\usage{
TLsort(usin)
}
\arguments{
\item{usin}{The community to be sorted.}
}
\value{
The community returned after sorting
}
\description{
Sorts the trophic levels from lowest to highest
}
\examples{
TLsort(intro_comm)
}
