% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{can_mutfeed}
\alias{can_mutfeed}
\title{A function that identifies cannibalism and mutual feeding}
\usage{
can_mutfeed(usin)
}
\arguments{
\item{usin}{The community in which to identify cannibalism and mutual feeding}
}
\value{
The community with mutual feeding added to the properties database.
}
\description{
A function that identifies cannibalism and mutual feeding
}
\examples{
intro_comm_mod = intro_comm
intro_comm_mod$imat["Predator", "Predator"] = 1
can_mutfeed(intro_comm_mod)
}
