% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{checkcomm}
\alias{checkcomm}
\title{A function check the community for errors before it is used in calculations.}
\usage{
checkcomm(usin, shuffleTL = FALSE, rmzeros = TRUE, verbose = TRUE)
}
\arguments{
\item{usin}{The community to check.}

\item{shuffleTL}{A Boolean stating whether the community should be sorted.}

\item{rmzeros}{A Boolean determining whether trophic species with zero biomass should be removed from the community.}

\item{verbose}{A Boolean. Do you want to see all the warnings?}
}
\value{
The checked community.
}
\description{
A function check the community for errors before it is used in calculations.
}
\examples{
checkcomm(intro_comm)
}
