% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whomineralizes.R
\name{whomineralizes}
\alias{whomineralizes}
\title{Direct and indirect contributions to mineralizations}
\usage{
whomineralizes(usin)
}
\arguments{
\item{usin}{The community in which we want to calculate mineralization rates.}
}
\value{
A table of mineralization rates.
}
\description{
Direct and indirect contributions to mineralizations
}
